+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/extension/ExtensionMethods.scala
@@ -41,6 +41,7 @@ object ExtensionMethods {
   val extensionMethodsHandlers: List[ExtensionMethodsHandler] = List(
     Cast,
     ArrayExt,
+    CollectionConversionExt,
   )
 
   def enrichWithExtensionMethods(set: ClassDefinitionSet): ClassDefinitionSet = {