+++ scenario-compiler/src/test/scala/pl/touk/nussknacker/engine/extension/ExtensionMethodsSpec.scala
@@ -27,11 +27,23 @@ class ExtensionMethodsSpec extends AnyFunSuite with Matchers {
     )
     val definitionsSet = ClassDefinitionSet(Set(stringDefinition, unknownDefinition))
 
-    ExtensionMethods.enrichWithExtensionMethods(
-      definitionsSet
-    ).classDefinitionsMap.map(e => e._1.getName -> e._2.methods.keys) shouldBe Map(
+    ExtensionMethods
+      .enrichWithExtensionMethods(
+        definitionsSet
+      )
+      .classDefinitionsMap
+      .map(e => e._1.getName -> e._2.methods.keys) shouldBe Map(
       "java.lang.String" -> Set("toUpperCase"),
-      "java.lang.Object" -> Set("toString", "canCastTo", "castTo", "castToOrNull"),
+      "java.lang.Object" -> Set(
+        "toString",
+        "canCastTo",
+        "castTo",
+        "castToOrNull",
+        "toMap",
+        "toMapOrNull",
+        "toList",
+        "toListOrNull"
+      ),
     )
   }